+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Started maintaining a changelog
 * `Config::comment_defaults` allows setting `//@` comments for all tests
 * `//~` comments can now specify just an error code or lint name, without any message. ERROR level is implied
+* `Revisioned::diagnostic_code_prefix` allows stripping a prefix of diagnostic codes to avoid having to repeat `clippy::` in all messages
 
 ### Fixed